service/requirements.txt
Flask==0.12.2
flask-restplus==0.10.1
python-box==3.1.1
